[CI skip] Reduced technical debt
