+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/AutomatedCraftingChamber.java
@@ -36,6 +36,13 @@ import me.mrCookieSlime.Slimefun.api.inventory.BlockMenuPreset;
 import me.mrCookieSlime.Slimefun.api.inventory.DirtyChestMenu;
 import me.mrCookieSlime.Slimefun.api.item_transport.ItemTransportFlow;
 
+/**
+ * This class needs to be rewritten VERY BADLY.
+ * But we should focus on rewriting the recipe system first.
+ * 
+ * @author TheBusyBiscuit
+ *
+ */
 public abstract class AutomatedCraftingChamber extends SlimefunItem implements InventoryBlock, EnergyNetComponent {
 
     private final int[] border = { 0, 1, 3, 4, 5, 7, 8, 13, 14, 15, 16, 17, 50, 51, 52, 53 };
@@ -233,11 +240,13 @@ public abstract class AutomatedCraftingChamber extends SlimefunItem implements I
             }
 
             ItemStack item = menu.getItemInSlot(getInputSlots()[j]);
+
             if (item != null && item.getAmount() == 1) {
-                if (craftLast)
+                if (craftLast) {
                     lastIteration = true;
-                else
+                } else {
                     return "";
+                }
             }
 
             builder.append(CustomItemSerializer.serialize(item, ItemFlag.MATERIAL, ItemFlag.ITEMMETA_DISPLAY_NAME, ItemFlag.ITEMMETA_LORE));
@@ -248,8 +257,9 @@ public abstract class AutomatedCraftingChamber extends SlimefunItem implements I
         // we're only executing the last possible shaped recipe
         // we don't want to allow this to be pressed instead of the default timer-based
         // execution to prevent abuse and auto clickers
-        if (craftLast && !lastIteration)
+        if (craftLast && !lastIteration) {
             return "";
+        }
 
         return builder.toString();
     }
