+++ src/main/java/me/mrCookieSlime/Slimefun/api/inventory/UniversalBlockMenu.java
@@ -33,8 +33,10 @@ public class UniversalBlockMenu extends DirtyChestMenu {
     }
 
     public void save() {
-        if (!isDirty())
+        if (!isDirty()) {
             return;
+        }
+
         // To force CS-CoreLib to build the Inventory
         this.getContents();
